Restored std::size_t workspace in PerfField

+++ projects/miopen/src/find_db.cpp
@@ -272,7 +272,7 @@ void FindDbRecord_t<TDb>::CopyTo(std::vector<Solution>& to) const
         const auto range = content->As<FindDbData>();
         std::transform(range.begin(), range.end(), std::back_inserter(to), [](const auto& pair) {
             return Solution{
-                solver::Id{pair.first}, pair.second.time, size_t(pair.second.workspace)};
+                solver::Id{pair.first}, pair.second.time, pair.second.workspace};
         });
     }
     else
